refactor: 전체 구조 리팩터링
