+++ src/main/java/domain/game/Game.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import domain.ladder.Line;
 import java.util.ArrayList;
